[Bug Fix] Removing b from plt.grid
The plt.grid(b=None) statement is causing the issue. The b argument in plt.grid() is used to specify whether to show the grid lines or not. When b is set to None, it tries to toggle the grid state, but Colab through this error `ValueError: keyword grid_b is not recognized; valid keywords`.